Move dialogs to presentation package
