+++ app/src/main/java/com/example/traders/presentation/dialogs/buyDialog/BuyDialogViewModel.kt
@@ -7,13 +7,13 @@ import androidx.lifecycle.ViewModelProvider
 import com.example.traders.BaseViewModel
 import com.example.traders.database.Crypto
 import com.example.traders.database.TransactionType
-import com.example.traders.dialogs.DialogValidation
-import com.example.traders.dialogs.DialogValidationMessage
-import com.example.traders.dialogs.validateChars
+import com.example.traders.presentation.dialogs.DialogValidation
+import com.example.traders.presentation.dialogs.DialogValidationMessage
 import com.example.traders.profile.portfolio.TransactionInfo
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.example.traders.database.FixedCryptoList
+import com.example.traders.presentation.dialogs.validateChars
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
