+++ app/src/main/java/com/example/traders/presentation/dialogs/confirmationDialog/ConfirmationDialogEvent.kt
@@ -1,4 +1,4 @@
-package com.example.traders.dialogs.confirmationDialog
+package com.example.traders.presentation.dialogs.confirmationDialog
 
 sealed class ConfirmationDialogEvent {
     object Dismiss: ConfirmationDialogEvent()
