+++ app/src/main/java/com/example/traders/presentation/dialogs/depositDialog/DepositDialogEvent.kt
@@ -1,4 +1,4 @@
-package com.example.traders.dialogs.depositDialog
+package com.example.traders.presentation.dialogs.depositDialog
 
 import com.example.traders.database.Transaction
 import java.math.BigDecimal
