+++ app/src/main/java/com/example/traders/presentation/dialogs/sellDialog/SellDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.example.traders.dialogs.sellDialog
+package com.example.traders.presentation.dialogs.sellDialog
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -13,7 +13,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.example.traders.R
 import com.example.traders.databinding.DialogFragmentSellBinding
-import com.example.traders.dialogs.DialogValidationMessage
+import com.example.traders.presentation.dialogs.DialogValidationMessage
 import com.example.traders.database.FixedCryptoList
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collect
