+++ app/src/main/java/com/example/traders/presentation/dialogs/sellDialog/SellState.kt
@@ -1,7 +1,7 @@
-package com.example.traders.dialogs.sellDialog
+package com.example.traders.presentation.dialogs.sellDialog
 
 import com.example.traders.database.Crypto
-import com.example.traders.dialogs.DialogValidationMessage
+import com.example.traders.presentation.dialogs.DialogValidationMessage
 import java.math.BigDecimal
 
 data class SellState(
