+++ app/src/main/java/com/example/traders/profile/history/HistoryFragment.kt
@@ -8,8 +8,8 @@ import androidx.fragment.app.viewModels
 import com.example.traders.BaseFragment
 import com.example.traders.database.Transaction
 import com.example.traders.databinding.FragmentHistoryBinding
-import com.example.traders.dialogs.confirmationDialog.ConfirmationDialogFragment
-import com.example.traders.dialogs.confirmationDialog.ConfirmationType
+import com.example.traders.presentation.dialogs.confirmationDialog.ConfirmationDialogFragment
+import com.example.traders.presentation.dialogs.confirmationDialog.ConfirmationType
 import com.example.traders.profile.adapters.HistoryListAdapter
 import dagger.hilt.android.AndroidEntryPoint
