+++ app/src/main/res/layout/dialog_confirmation.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:padding="20dp"
-    tools:context=".dialogs.confirmationDialog.ConfirmationDialogFragment">
+    tools:context=".presentation.dialogs.confirmationDialog.ConfirmationDialogFragment">
 
     <TextView
         android:id="@+id/confirmation_msg"
