+++ app/src/test/java/com/example/traders/dialogs/buyDialog/BuyDialogViewModelTest.kt
@@ -3,7 +3,7 @@ package com.example.traders.dialogs.buyDialog
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.cash.turbine.test
 import com.example.traders.database.Crypto
-import com.example.traders.dialogs.DialogValidation
+import com.example.traders.presentation.dialogs.DialogValidation
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.example.traders.TestDispatcherRule
