+++ app/src/test/java/com/example/traders/dialogs/depositDialog/DepositViewModelTest.kt
@@ -2,9 +2,11 @@ package com.example.traders.dialogs.depositDialog
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.cash.turbine.test
-import com.example.traders.dialogs.DialogValidation
+import com.example.traders.presentation.dialogs.DialogValidation
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.TestDispatcherRule
+import com.example.traders.presentation.dialogs.depositDialog.DepositDialogEvent
+import com.example.traders.presentation.dialogs.depositDialog.DepositViewModel
 import com.nhaarman.mockitokotlin2.mock
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.launch
